General update. Major logic layer update.

+++ logic/logicfacade.py
@@ -14,6 +14,7 @@ from validationlib import (idChecksum,
                            isCountryValid)
 from utilitieslib import classes, getPlural
 from logicexceptions import *
+
 from persistence.persistence_manager import fileDataManager, countryDataManager
 
 countries = countryDataManager.get()
@@ -21,6 +22,8 @@ countries = countryDataManager.get()
 
 class LogicFacade(ABC):
     '''
+        status = WIP (75%)
+
         Static class that defines static methods meant to be called from API.
 
         Each method handles a particular HTTP request.
@@ -42,27 +45,22 @@ class LogicFacade(ABC):
 
         --Exceptions--
         from logicexceptions.py:
-        TypeError:
+        TypeError, ValueError:
             getByType()
             getByID()
             createObjectByJson()
             updatedByID()
             deleteByID()
-        IDChecksumError:
-            getByID()
-            deleteByID()
-            updateByID()
-            createObjectByJson()
-        InvalidCountryCode, CountryNotFoundError:
+        CountryNotFoundError:
             updateByID()
             createObjectByJson()
             getCountry()
             getCountryCities()
-        EmailDuplicated, IDNotFound, ValueError, TypeError:
+        EmailDuplicated:
             updateByID()
             createObjectByJson()
         from db.dbexceptions.py:
-        IDNotFound:
+        IDNotFoundError:
             getByID()
             deleteByID()
             updateByID()
@@ -77,21 +75,26 @@ class LogicFacade(ABC):
         return fileDataManager.getAll(clsPlural)
 
     @staticmethod
-    def getByID(id: str, cls: str) -> str:
+    def getByID(id, cls: str) -> str:
+        if not isinstance(id, str):
+            raise TypeError("id must be a string")
         if not idChecksum(id):
-            raise IDChecksumError("invalid id")
+            raise ValueError("invalid id")
         clsPlural = getPlural(cls)
         return fileDataManager.get(id, clsPlural)
 
     @staticmethod
     def deleteByID(id: str, cls: str) -> None:
+        if not isinstance(id, str):
+            raise TypeError("id must be a string")
         if not idChecksum(id):
-            raise IDChecksumError("invalid id")
+            raise ValueError("invalid id")
         clsPlural = getPlural(cls)
         return fileDataManager.delete(id, clsPlural)
 
     @staticmethod
     def updateByID(id: str, cls: str, data: str) -> None:
+        # id check made within the class __init__
         for c in classes:
             # c[0]: singular str, c[1]: plural str, c[2]: class
             if cls == c[0]:
@@ -101,17 +104,21 @@ class LogicFacade(ABC):
 
     @staticmethod
     def createObjectByJson(cls: str, data: str) -> None:
+        # id check made within the class __init__
         for c in classes:
             # c[0]: singular str, c[1]: plural str, c[2]: class
             if cls == c[0]:
-                new = c[2].createFromJson(data)
+                new = c[2].createFromJson(json.loads(data))
                 clsPlural = c[1]
-        return fileDataManager.save(clsPlural, new)
+                fileDataManager.save(clsPlural, new)
+        raise TypeError("invalid class")
 
     @staticmethod
-    def getCountry(code: str) -> str:
+    def getCountry(code) -> str:
+        if not isinstance(code, str):
+            raise TypeError("country code must be a string")
         if not isCountryValid(code):
-            raise InvalidCountryCode("country code is not valid")
+            raise ValueError("country code is not valid")
         for country in countries:
             if country.code == code:
                 return country
@@ -119,11 +126,12 @@ class LogicFacade(ABC):
 
     @staticmethod
     def getContryCities(code: str) -> str:
+        if not isinstance(code, str):
+            raise TypeError("country code must be a string")
         if not isCountryValid(code):
-            raise InvalidCountryCode("country code is not valid")
+            raise ValueError("country code is not valid")
         for country in countries:
             if country.code == code:
-                # code corresponds to an existing country code
                 return fileDataManager.getAllWithProperty(
                         "cities", "country_code", code)
         raise CountryNotFoundError("country does not exist")
